+++ next.config.js
@@ -1,6 +1,7 @@
 /** @type {import('next').NextConfig} */
 const withPWA = require("next-pwa")({
   dest: "public",
+  disable: process.env.NODE_ENV === "development",
 });
 
 const nextConfig = {
