+++ src/controller/postController/useUpdatePostController.ts
@@ -1,9 +1,6 @@
 import { postApi } from "@/api/postApi";
-import { QUERY_KEY } from "@/config/queyKeyConfig";
-import { postAtom } from "@/stateStore/postAtom";
-import { Post } from "@/types/model/post";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { useSetRecoilState } from "recoil";
+import useMutationService from "@/service/postService/useMutationService";
+import { useMutation } from "@tanstack/react-query";
 
 interface MutateParameter {
   title: string;
@@ -11,27 +8,18 @@ interface MutateParameter {
 }
 
 function useUpdatePostController(postId: number) {
-  const queryClient = useQueryClient();
-  const setPost = useSetRecoilState(postAtom);
+  const { handleOptimisticUpdate, replaceFreshTitleCacheForPagination, handleRollback } = useMutationService(postId);
 
   return useMutation({
     mutationFn: ({ title, contents }: MutateParameter) => postApi.update(postId, title, contents),
-    onMutate: ({ title, contents }) => {
-      const prevPost = queryClient.getQueryData<Post>([QUERY_KEY.post, postId]);
-      setPost((prev) => ({ ...prev, title, contents }));
-      queryClient.setQueryData([QUERY_KEY.post, postId], {
-        ...prevPost,
-        title,
-        contents,
-      });
-      return { prevPost };
+    onMutate: (variables) => {
+      return handleOptimisticUpdate(variables);
+    },
+    onSuccess(_, variables) {
+      replaceFreshTitleCacheForPagination(variables);
     },
     onError: (error, _, context) => {
-      if (!context) return;
-      if (context?.prevPost) {
-        queryClient.setQueryData([QUERY_KEY.post, postId], context.prevPost);
-        setPost(context.prevPost);
-      }
+      handleRollback(context);
     },
   });
 }
